@@ -10,12 +10,13 @@ dependencies:
   flutter_bloc: 3.2.0
   equatable: ^1.1.0
   built_value: 6.7.1
-  onesignal_flutter: ^2.0.0
+  onesignal_flutter: ^2.6.1
   json_annotation:
   http: ^0.12.1
   flutter:
     sdk: flutter
-  firebase_auth: ^0.14.0+5
+  firebase_core: ^0.5.0
+  firebase_auth: ^0.18.0+1
   basic_utils: ^2.5.3
 
   cupertino_icons: ^0.1.2
