@@ -1,11 +1,13 @@
+import 'package:GTUBT/models/user.dart';
 import 'package:GTUBT/service/authentication.dart';
+import 'package:GTUBT/service/user.dart';
 import 'package:GTUBT/ui/blocs/register_bloc/bloc.dart';
 import 'package:GTUBT/ui/utils/validators.dart';
 import 'package:bloc/bloc.dart';
 
 class RegisterBloc extends Bloc<RegisterEvent, RegisterState> {
   final AuthService _authService = AuthService();
-
+  final UserService _userService = UserService();
   RegisterBloc();
 
   @override
@@ -17,14 +19,14 @@ class RegisterBloc extends Bloc<RegisterEvent, RegisterState> {
       yield state.update(isEmailValid: Validators.isValidEmail(event.email));
     } else if (event is NameChanged) {
       yield state.update(isNameValid: Validators.isValidString(event.name));
-    } else if (event is SurnameChanged) {
+    } else if (event is LastnameChanged) {
       yield state.update(
-          isSurnameValid: Validators.isValidString(event.surname));
+          isLastnameValid: Validators.isValidString(event.lastname));
     } else if (event is PasswordChanged) {
       yield state.update(
           isPasswordValid: Validators.isValidPassword(event.password));
     } else if (event is ClassChanged) {
-      yield state.update(isGradeValid: Validators.isValidDigit(event.grade));
+      yield state.update(isGradeValid: Validators.isValidDigit(event.year));
     } else if (event is PhoneNumberChanged) {
       yield state.update(
           isPhoneNumberValid: Validators.isValidPhoneNumber(event.phoneNumber));
@@ -44,22 +46,23 @@ class RegisterBloc extends Bloc<RegisterEvent, RegisterState> {
     yield RegisterState.loading();
 
     try {
-//      TODO: make into model
-      Map<String, dynamic> data = {
-        'email': event.email,
-        'password': event.password,
-        'name': event.name,
-        'surname': event.surname,
-        'grade': event.grade,
-        'department': event.department,
-        'phoneNumber': event.phoneNumber,
-        'studentNumber': event.studentNumber
-      };
+      User user = User(
+          email: event.email,
+          name: event.name,
+          lastname: event.lastname,
+          year: int.parse(event.year),
+          department: Department.cse,
+          phone: event.phoneNumber,
+          studentId: event.studentNumber);
 
-      await _authService.signUp(data);
-      // TODO: create user if signUp is successful
+      await _authService.signUp({
+        'email': user.email,
+        'password': event.password,
+      });
+      await _userService.post(user);
       yield RegisterState.success();
     } catch (error) {
+      print(error);
       yield RegisterState.failure();
     }
   }
