@@ -3,6 +3,7 @@ import 'package:GTUBT/ui/blocs/authentication_bloc/authentication_state.dart';
 import 'package:GTUBT/ui/blocs/page_bloc/bloc.dart';
 import 'package:GTUBT/ui/blocs/user_bloc/bloc.dart';
 import 'package:GTUBT/ui/blocs/user_bloc/user_bloc.dart';
+import 'package:GTUBT/ui/components/hamburgerMenu.dart';
 import 'package:GTUBT/ui/routes.dart';
 import 'package:GTUBT/ui/style/color_sets.dart';
 import 'package:flutter/material.dart';
@@ -36,60 +37,6 @@ class _MainPageState extends State<MainPage> {
     BlocProvider.of<UserBloc>(context).add(ToggleEditMode());
   }
 
-  List<Widget> _getDrawers(
-      BuildContext context, AuthenticationState authState) {
-    var drawerChilds = <Widget>[];
-    if (authState is AuthenticationAuthenticated) {
-      drawerChilds.add(InkWell(
-        child: Container(
-          margin: EdgeInsets.symmetric(
-            vertical: 10,
-          ),
-          height: 50,
-          color: ColorSets.profilePageThemeColor,
-          child: const Center(child: Text('LOGOUT')),
-        ),
-        onTap: () {
-          // TODO: Add logout function here
-          BlocProvider.of<PageBloc>(context).add(
-            PageChanged(
-              context: context,
-              routeName: LOGIN_URL,
-            ),
-          );
-        },
-      ));
-    } else {
-      drawerChilds.add(InkWell(
-        child: Container(
-          margin: EdgeInsets.symmetric(
-            vertical: 10,
-          ),
-          height: 50,
-          color: ColorSets.profilePageThemeColor,
-          child: const Center(child: Text('LOGIN')),
-        ),
-        onTap: () {
-          BlocProvider.of<PageBloc>(context).add(
-            PageChanged(
-              context: context,
-              routeName: LOGIN_URL,
-            ),
-          );
-        },
-      ));
-    }
-    drawerChilds.add(Container(
-      margin: EdgeInsets.symmetric(
-        vertical: 10,
-      ),
-      height: 50,
-      color: ColorSets.profilePageThemeColor,
-      child: const Center(child: Text('SETTINGS')),
-    ));
-    return drawerChilds;
-  }
-
   @override
   Widget build(BuildContext context) {
     var actions = <Widget>[];
@@ -111,45 +58,32 @@ class _MainPageState extends State<MainPage> {
       child: BlocBuilder<PageBloc, PageState>(
         builder: (context, state) {
           return BlocBuilder<AuthenticationBloc, AuthenticationState>(
-            builder: (context, authState) {
-              var drawers = _getDrawers(context, authState);
-              return Scaffold(
-                appBar: AppBar(
-                  backgroundColor: ColorSets.barBackgroundColor,
-                  title: Text(Routes.bodyTitle[state.currentPage]),
-                  actions: actions,
+              builder: (context, authState) {
+            return Scaffold(
+              appBar: AppBar(
+                backgroundColor: ColorSets.barBackgroundColor,
+                title: Text(Routes.bodyTitle[state.currentPage]),
+                actions: actions,
+              ),
+              bottomNavigationBar: BottomNavigationBar(
+                unselectedItemColor: ColorSets.unselectedBarItemColor,
+                selectedIconTheme: IconThemeData(
+                  color: ColorSets.selectedBarItemColor,
                 ),
-                bottomNavigationBar: BottomNavigationBar(
-                  unselectedItemColor: ColorSets.unselectedBarItemColor,
-                  selectedIconTheme: IconThemeData(
-                    color: ColorSets.selectedBarItemColor,
-                  ),
-                  unselectedIconTheme: IconThemeData(
-                    color: ColorSets.unselectedBarItemColor,
-                  ),
-                  currentIndex: state.currentPage,
-                  backgroundColor: ColorSets.barBackgroundColor,
-                  onTap: _onNavigation,
-                  items: authState is AuthenticationAuthenticated
-                      ? Routes.navListLoggedIn
-                      : Routes.navList,
+                unselectedIconTheme: IconThemeData(
+                  color: ColorSets.unselectedBarItemColor,
                 ),
-                drawer: Drawer(
-                  child: Container(
-                    color: ColorSets.barBackgroundColor,
-                    alignment: Alignment.center,
-                    child: ListView(
-                      padding: EdgeInsets.zero,
-                      shrinkWrap: true,
-                      children: drawers,
-                    ),
-                  ),
-                ),
-                body: Routes.bodyList[state.currentPage],
-                // This trailing comma makes auto-formatting nicer for build methods.
-              );
-            },
-          );
+                currentIndex: state.currentPage,
+                backgroundColor: ColorSets.barBackgroundColor,
+                onTap: _onNavigation,
+                items: authState is AuthenticationAuthenticated
+                    ? Routes.navListLoggedIn
+                    : Routes.navList,
+              ),
+              drawer: HamburgerMenuComponents(),
+              body: Routes.bodyList[state.currentPage],
+            );
+          });
         },
       ),
     );
